# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse:
    properties:
      code:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode'
      error:
        type: string
    type: object
...
  github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode:
    enum:
    - INVALID_CREDENTIALS
    - UNAUTHORIZED
    - FORBIDDEN
    - EMAIL_NOT_VERIFIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_ALREADY_EXISTS
    - INVALID_REQUEST
    - VALIDATION_FAILED
    - MISSING_REQUIRED_FIELD
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - OPERATION_FAILED
    - EMAIL_IS_ALREADY_IN_USE
    - EMAIL_IS_ALREADY_VERIFIED
    type: string
    x-enum-varnames:
    - ErrCodeInvalidCredentials
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeEmailNotVerified
    - ErrCodeResourceNotFound
    - ErrCodeResourceAlreadyExists
    - ErrCodeInvalidRequest
    - ErrCodeValidationFailed
    - ErrCodeMissingRequiredField
    - ErrCodeInternalServer
    - ErrCodeDatabaseError
    - ErrCodeOperationFailed
    - ErrCodeEmailIsAlreadyInUse
    - ErrCodeEmailIsAlreadyVerified
...
  github_com_kotolino_lawyer_internal_handlers_responses.LawyerResponse:
    properties:
      address:
        type: string
      bar_association:
        type: string
      bar_number:
        type: string
      bio:
        type: string
      email:
        type: string
      experience_years:
        type: integer
      full_name:
        type: string
      id:
        type: integer
      languages:
        items:
          type: string
        type: array
      office_name:
        type: string
      profile_image:
        type: string
      rating:
        type: number
      review_count:
        type: integer
      specialties:
        items:
          type: string
        type: array
      user_id:
        type: integer
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginationResponse'
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.PaginationResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kotolino_lawyer_internal_models.Answer:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_accepted:
        type: boolean
      is_hidden:
        type: boolean
      lawyer:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.Lawyer'
      lawyer_id:
        type: integer
      question_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_kotolino_lawyer_internal_models.Availability:
    additionalProperties: true
    type: object
...
  github_com_kotolino_lawyer_internal_models.Lawyer:
    properties:
      achievements:
        type: string
      active:
        type: boolean
      address:
        type: string
      affiliation:
        type: string
      areas_of_expertise:
        type: string
      availability:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.Availability'
      average_rating:
        type: number
      bar_association:
        type: string
      bar_number:
        type: string
      bio:
        type: string
      certification_document_path:
        type: string
      certifications:
        type: string
      created_at:
        type: string
      email:
        type: string
      experience_years:
        type: integer
      fax_number:
        type: string
      full_name:
        type: string
      id:
        type: integer
      languages:
        items:
          type: string
        type: array
      lawyer_registration_number:
        type: string
      notes:
        type: string
      office_name:
        type: string
      phone:
        type: string
      phone_number:
        type: string
      profile_text:
        type: string
      rating:
        type: integer
      review_count:
        description: Additional calculated fields that don't exist in the database
        type: integer
      specialties:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.User'
      user_id:
        type: integer
    type: object
...
  github_com_kotolino_lawyer_internal_models.User:
    properties:
      address:
        type: string
      age_group:
        type: string
      birth_date:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      nickname:
        type: string
      notes:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      prefecture:
        type: string
      profile_image:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.AnswerResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_accepted:
        type: boolean
      lawyer:
        allOf:
        - $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.LawyerResponse'
        description: New field for the lawyer object
      lawyer_id:
        type: integer
      question_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /answers/{id}:
    delete:
      description: Deletes an answer (must be the author)
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Answer deleted
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid answer ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden - must be a lawyer
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Delete answer
      tags:
      - answers
    get:
      description: Retrieves an answer by its ID
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.Answer'
        "400":
          description: Invalid answer ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "404":
          description: Answer not found
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get answer by ID
      tags:
      - answers
    put:
      consumes:
      - application/json
      description: Updates an existing answer
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      - description: New answer content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AnswerResponse'
        "400":
          description: Invalid answer ID or request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden - must be the answer author
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Update answer
      tags:
      - answers
  /answers/{id}/accept:
    put:
      description: Marks an answer as accepted by the question owner
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Answer accepted
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid answer ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Accept answer
      tags:
      - answers
...
  /lawyers/{id}/answers:
    get:
      description: Retrieves all answers by a specific lawyer with pagination
      parameters:
      - description: Lawyer ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.AnswerResponse'
                  type: array
              type: object
        "400":
          description: Invalid lawyer ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get lawyer answers
      tags:
      - answers
      - lawyers
...
  /me/answers:
    get:
      description: Retrieves all answers by the currently authenticated lawyer with
        pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.AnswerResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden - must be a lawyer
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get my answers
      tags:
      - answers
...
  /questions/{id}/answers:
    get:
      description: Retrieves all answers for a specific question with pagination
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.AnswerResponse'
                  type: array
              type: object
        "400":
          description: Invalid question ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get answers by question
      tags:
      - answers
...
  /questions/{question_id}/answers:
    post:
      consumes:
      - application/json
      description: Creates a new answer to a question
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: integer
      - description: Answer content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AnswerResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden - must be a lawyer
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Create answer
      tags:
      - answers
...
swagger: "2.0"